requirements.txt
Flask==3.1.3
pandas==3.0.2
scikit-learn==1.8.0
transformers==5.4.0
torch==2.11.0
tqdm==4.67.3
sentence-transformers
faiss-cpu
numpy
